+++ CHANGELOG.md
@@ -6,9 +6,10 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 ## 0.9.1
 
-2020 Aug 4 [in progress]
+2020 Aug 5 [in progress]
 
 *   Added "system", "dark", and "light" theme options #72.
+*   Added frequency analysis that computes the Welch periodogram.
 
 
 ## 0.9.0